Added sigma.classes.graph.read method
This methods takes a graph object as input: An object with an array of valid node objects and an array of valid edge objects, associated respectively to the keys "nodes" and "edges".

Also added the related unit tests, and called this method in sigma constructor if a graph is given in the configuration.