Improved quad behavior before indexation
I met some troubles sometimes calling the retrieve methods of the quad ("point" and "area") before the indexation, so I just initialized the _tree attribute as null when instanciated, and checked that _tree is not null before calling those methods. If its null, then the quad will now just return an empty array.