Fixed rendering process when autoRescale is false
When the setting "autoRescale" is set to false, the "rescale" middleware is not called, and the camera was supposed to guess what reading prefix to take.
So, from now, when the "autoRescale" is false, the "copy" middleware is called instead. Also, no more guessing from the camera: The camera does not care anymore about the "autoRescale" setting value.