Added a custom node renderer example
This is a basic example on how to develop a custom node renderer. In this example, the renderer will display an image clipped in a disc, with a border colored according the node's "color" value.

If a node as the value "image" to its attribute "type", then it will displayed with the node renderer "sigma.canvas.nodes.image", with the url being its "url" value.

IMPORTANT: This node renderer just works with the canvas renderer. If you do want to display images with the WebGL renderer, you will have to develop a specific WebGL node renderer.