Added a custom edge renderer example
This is a basic example on how to develop a custom edge renderer. In this example, the renderer will display the edge as a path made of two segments, an horizontal one and a vertical one. To make it just a bit "more readable", I also used the FA2 plugin (but it is actually  still not readable).

IMPORTANT: This edge renderer just works with the canvas renderer. If you do want to display images with the WebGL renderer, you will have to develop a specific WebGL edge renderer.