JSON parser plugin: sigma.parsers.json
This new plugins is basically just an ajax tool that loads a JSON graph and fills or creates an instance with the graph. The main idea is to avoid using jQuery just to load a file, since sigma already deals by itself with resizing, DOM, etc...

It comes with its example, based on the same dataset than the GEXF example.