Fine tuning on existing captors
The issue was visible on the examples/force.html sample: If the mouse was over a node, the hovered node was staying hovered as long as the mouse does not move.

The problem was that I removed earlier the event binding to the render event. It is restored, and the touch captor il now dispatching the "mouseMove" event, to fix the bug I tried to fix when I removed the binding...