Fixed #189: Added doubleClick events
Four new events are available: "doubleClick", "doubleClickStage", "doubleClickNode" and "doubleClickNodes".
The events.html example has also been modified to display "doubleClickStage" and "doubleClickNode".

Finally, a new setting, "doubleClickEnabled", defines if the graph must zoom in when the user doubleclicks.
Its default value is "true", but it is adviced to disable it when you want to catch the new double-click related events.