Fixed buggy drag behaviour when container is offset.
If _container is offset from the top-left corner of the page, it is very easy for local/global mouse position errors to creep into calculations. In particular, when dragging a node it hovers a fixed distance away from the mouse. This more robust calculation fixes these issues (including padding which is independent of the bounding rect), and allows the container to be placed anywhere with any offset/margin/padding, and it need not be position: absolute.