Related to #263: Restored renderers.def as backup
In d74879... I replaced calling the "def" renderers if not renderer has been found, by using the defaultNodeType / defaultEdgeType for finding the renderer.
Since it caused too much issues with labels renderers, I restored using renderers.def as backup, if no renderer has been found with the node / edge type or the defaultNodeType / defaultEdgeType value.

Later, it will be possible to select which node / edge / label renderer to use from a function, when #368 will be fixed.