Fix browser tab crashing on attempt to download
I was trying to save my sigma instance as a `.gexf` file by using the `sigma.exporters.gexf` plugin. But, if my created file is quite large, say around 4MB, browser tab crashes.

I found the problem is in the download function, where the file is generated as a dataURI which would obviously be very big for a file of that size.

So I changed the download method to use a blob to create the file instead of dataURI, and that seems to work.