Moves position translates business to renderer
It was previously held by the camera, but induced some problems, since
positions need viewport dimensions to be translated while the camera
does not actually know the viewport.

And since in the future translations might need the nodes extent in
addition, it is a good time to move that business to the renderer.