#1132 - Fixes hideEdgesOnMove behavior
The issue was that, when using the `hideEdgesOnMove` setting, when the
user drags, waits, and then drops the graph, the edges were not rendered
at all.

The reason is that, basically, because the user waits, the camera state
will be unchanged once the user dropsthe graph, so there will be no new
rendering to display the edges.

The solution is to add a `renderer.refresh()` call "for next frame", in
the `mouseup` handler. Sometimes this refresh will be too much, but it
is not too much of an issue, and it will solve this bug.