#1135 - Fixes enterNode for extreme nodes
An early exit in `QuadTree#point` was causing the bug (when the mouse is
not over the rectangle that contains all node centers, it was not
returning any node candidate).

The solution is to always search the quadtree even when the mouse is
far, and also to always return all unindexed nodes as well, to handle
some rare edge cases.