Drafts new website
Since the new website is a single page, the whole Next.js work
previously done has been erased. Also, since we found a way to easily
open a CodeSandbox on an existing and up-to-date sigma example, there is
no more need to build the examples and put them online.

This website is not fully done yet though:
- Main showcase example must be developed
- The "In the wild" section must be implemented
- The Cookbook section probably needs more content