Drafts #getEdgeAtPoint sigma method
This method will later allow to deal with edge events.
The current version has acceptable early exitsi
- it first check if the pixel is colored in the edge layer asdasd asdasd
- it then checks for each edge if the node is "roughly" between the
extremities
- finally, it really checks the exact collision

The main issue is that final exact checks now translates for each edge
both extremities coordinates to the viewport ref, instead of just
translating once the mouse coordinates to the graph ref.
The reason is that I couldn't yet translate the edge thickness to the
graph ref.