#1186 - Separates "mousemovebody" and "mousemove"
The "mousemove" event is only dispatched by the mousecaptor when the
mouse is exactly hovering the sigma instance container.

The "mousemovebody" event is always dispatched, so that it is possible
to properly implement drag-and-dropping, even while the mouse is out of
the container.

This commit also fixes the mouse-manipulations example, btw.