Refacto refresh function
Function refresh can updates all dataCache, update the programs data and call a render if a partial graph is provided.
It bypass all actions related to nodes coordinates (like quadtree, labelGrid) if layoutUnchanged is at true..
Moreover, you can also specify if it should do a render or schedule it.