[sigma] Fixes #1399
This commit adds the devicePixelRatio into consideration when calling
getPixelColor as suggested by foxton9 (thanks a lot!), and also fixes
the texture size.

Without fixing the texture size, the first fix would work, but only on a
fraction of the stage, with a devicePixelRatio different than 1 (for
instance, with a devicePixelRatio of 2, only a fourth of the screen
would be interactive...).