[storybook] Updates custom-rendering example
Since @sigma/node-border is now working, to get a custom node program
that is not already implemented elsewhere for the custom-rendering
story, the local node.border.ts program is replaced by a new
node.gradient.ts program.

It's basically the same, but with the fragment shader modified so that
it draws a gradient instead of a border.