[sigma] Makes arrow heads size customizable
There is one thing to be aware of though:
To facilitate displaying straight arrows and curved arrows in the same
graph (to have a better parallel edges rendering for instance), this
commit changes the default arrows shape. It is still possible to render
arrows as they were before, though, by creating a custom
edgeArrowProgram.

Details:
- Replaces EdgeArrowHeadProgram by createEdgeArrowHeadProgram, that
  accepts custom lengthToThicknessRatio and widenessToThicknessRatio
- Replaces EdgeClampedProgram by createEdgeClampedProgram, that accepts
  a custom lengthToThicknessRatio
- Adds new factory createEdgeArrowProgram, and EdgeArrowProgram is built
  using this factory